requirements.txt
setuptools
numpy
tensorflow
scikit-learn
keras
pandas
matplotlib
seaborn
scipy
biopython
openpyxl